已經完成版本1.2beta, 主要的作爲是把powdertoy的文字繪製部分移植到3body上, 詳細見README文檔.

+++ mathlib.h
@@ -12,6 +12,6 @@ double direction(const Object& theObj, const Object& target);
 Vector gravitation(const Object& theObj, const Object& target, double G);
 
 void drawCircle(float x, float y, double radius,
-                float r, float g, float b, int div_parts);
+                int r, int g, int b, int div_parts);
 
 #endif /* __MATHLIB_H__ */
